Reorganize reinforcetrader modules for less imports
