+++ reinforcetrader/dqn_agent.py
@@ -1,0 +1,35 @@
+import keras
+from keras import Input, layers, optimizers
+
+@keras.saving.register_keras_serializable()
+# Define DQN Model Architecture
+class DualBranchDQN(keras.Model):
+    def __init__(self, motif_state_size: tuple[int, int], context_state_size: int, action_size: int):
+        super().__init__()
+
+        # Motif Branch for finding candle patterns using Conv1D
+        motif_input = Input(shape=motif_state_size, name="Motif Input")
+        motif_conv_layer = layers.Conv1D(32, 3, padding="same", activation="relu")(motif_input)
+        motif_conv_layer = layers.Conv1D(32, 3, padding="same", activation="relu")(motif_conv_layer)
+        motif_out = layers.GlobalAveragePooling1D(name="Motif Output")(motif_conv_layer)
+
+        # Context Branch for finding regimes
+        context_input = Input(shape=(context_state_size,), name="Context Input")
+        context_hid_layer = layers.Dense(64, activation="relu")(context_input)
+        context_out = layers.Dense(32, activation="relu", name="Context Output")(context_hid_layer)
+        
+        # Late fusion of both the motif and context branches
+        fused_branch = layers.Concatenate(name="Late Fusion")([motif_out, context_out])
+        fused_branch = layers.Dense(128, activation="relu")(fused_branch)
+        fused_branch = layers.Dense(64, activation="relu")(fused_branch)
+        fused_branch = layers.LayerNormalization()(fused_branch)
+        Q = layers.Dense(action_size, name="Q Values")(fused_branch)
+
+        self._model = keras.Model(inputs=[motif_input, context_input], outputs=Q, name="DualBranchDQN")
+        self._model.compile(loss="mse", optimizer=optimizers.Adam(1e-3))
+
+    def get_model(self):
+        return self._model
+		
+class RLAgent:
+    pass
